--- goapp/go.mod
@@ -1,7 +1,0 @@
-module sleeperpy/goapp
-
-go 1.21
-
-require (
-	github.com/patrickmn/go-cache v2.1.0+incompatible
-)
