--- goapp/templates/index.html
@@ -1,83 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>SleeperPy Fantasy Tiers</title>
-    <script src="https://unpkg.com/htmx.org@1.9.10"></script>
-    <link rel="stylesheet" href="/static/main.css">
-    <link rel="icon" type="image/svg+xml" href="/static/favicon.svg">
-</head>
-<body>
-        <header class="app-header">
-            <h1>SleeperPy (Go + HTMX)</h1>
-        </header>
-            <div class="container">
-                        <div class="app-description">
-                        <p><b>SleeperPy Fantasy Tiers</b> helps you set your best fantasy football lineup using <a href="https://www.borischen.co/" target="_blank" rel="noopener">Boris Chen's</a> weekly tiers.</p>
-                        <p>Enter your <b>Sleeper username</b> (<a href="https://support.sleeper.com/en/articles/1876011-account-settings" target="_blank" rel="noopener">how to find it</a>) to see your team, top free agents, and upgrade suggestions.</p>
-                        </div>
-        <form id="userform" method="get" action="/lookup">
-            <input id="inputButton" type="text" name="username" required placeholder="Sleeper Username">
-            <button id="generateTiers" type="submit">Show My Tiers</button>
-        </form>
-        <div id="loading" style="display:none;text-align:center;margin-top:1em;">Loading...</div>
-        <div id="output"></div>
-        <script>
-        document.body.addEventListener('htmx:configRequest', function(evt) {
-            document.getElementById('loading').style.display = 'block';
-        });
-        document.body.addEventListener('htmx:afterSwap', function(evt) {
-            document.getElementById('loading').style.display = 'none';
-        });
-        document.body.addEventListener('htmx:responseError', function(evt) {
-            document.getElementById('loading').style.display = 'none';
-        });
-        </script>
-    </div>
-<style>
-    .app-description {
-        background: #232c41;
-        color: #eaf0fa;
-        border-radius: 10px;
-        box-shadow: 0 2px 12px rgba(30,38,54,0.10);
-        padding: 22px 18px 18px 18px;
-        margin-bottom: 24px;
-        font-size: 1.13em;
-        max-width: 700px;
-        margin-left: auto;
-        margin-right: auto;
-    }
-    .app-description ul {
-        margin: 12px 0 0 0;
-        padding-left: 22px;
-    }
-    .app-description li {
-        margin-bottom: 6px;
-    }
-    .app-header {
-        width: 100%;
-        background: linear-gradient(90deg, #232c41 0%, #3a6ee8 100%);
-        padding: 32px 0 18px 0;
-        text-align: center;
-        margin-bottom: 18px;
-        box-shadow: 0 2px 12px rgba(30,38,54,0.10);
-    }
-    .app-header h1 {
-        color: #eaf0fa;
-        font-family: 'Raleway', Arial, sans-serif;
-        font-size: 2.3em;
-        font-weight: 800;
-        letter-spacing: 1px;
-        margin: 0;
-        text-shadow: 0 2px 8px rgba(30,38,54,0.18);
-    }
-</style>
-<footer style="margin-top:2em; text-align:center; font-size:1em; color:#888;">
-    <div style="margin-bottom:0.5em;">
-        Tiers powered by <a href="https://www.borischen.co/" target="_blank" rel="noopener">Boris Chen</a> |
-        <a href="https://github.com/wbollock/sleeperPy" target="_blank" rel="noopener">GitHub Repo</a>
-        <span style="margin-left:1em; color:#666;">If you like this project, please <a href="https://github.com/wbollock/sleeperPy" target="_blank" rel="noopener">drop a ⭐ on GitHub</a>!</span>
-    </div>
-</footer>
-</body>
-</html>
