--- tiers_test.html
@@ -1,66 +1,0 @@
-<!DOCTYPE html>
-<html>
-<body>
-    <style>
-        table, th, td {
-  border: 1px solid black;
-  border-collapse: collapse;
-        }
-    </style>
-    <h1>SleeperPy: Boris Chen Tiers for Sleeper Leagues</h1>
-    <p>Note:</p>
-    <ul>
-        <li>[X] indicates the tier of the player. Lower is better.</li>
-    </ul>
-    <h3>Username: puffplants</h3>
-
-    <table style="width:20%;">
-        <tr>
-          <th colspan="2">League: Duvalll | Scoring Type: PPR</th>
-        </tr>
-        <tr>
-            <th colspan="2">Starters</th>
-        </tr>
-        <tr>
-          <th>QB</th>
-          <th>Tier</th>
-        </tr>
-        <tr>
-            <td>Bill Gates</td>
-            <td>1</td>
-        </tr>
-        <tr>
-            <th>RB</th>
-            <th>Tier</th>
-          </tr>
-        <tr>
-            <td>Ingram</td>
-            <td>2</td>
-        </tr>
-        <tr>
-            <td>Joe Mixon</td>
-            <td>3</td>
-        </tr>
-        <tr>
-            <th colspan="2">Average Tier: 1.5</th>
-        </tr>
-        <tr>
-            <th colspan="2">Bench</th>
-        </tr>
-        <tr>
-            <th>RB</th>
-            <th>Tier</th>
-          </tr>
-        <tr>
-            <td>Robinson</td>
-            <td>8</td>
-        </tr>
-      </table>
-      
-    
-</body>
-</html>
-
-    
-
-
